requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
python-dotenv==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1
lxml==4.9.4
